Fix UART baudrate
The UART baudrate setup did not work as expected when using an 8 MHz clock. Setting the baudrate to 57600 would result in an actual baudrate of 9600.

In order to fix this behaviour, I'm changing the internal clock to 48 MHz using the free-dap's samd21 system initialization routine. This also significantly increases the read and write speed of SWD transfers.